Do not add phony entry to the end of mounted filesystems
If output of mount ends without a newline, phony entry was added
to list of mounted filesystems with options inherited from previous
filesystem and mountpoint set to empty string.

if last filesystem in output of mount was NFS one it resulted in errorneously
treating all filesystems as remote ones due to separate bug in filesystem
matching.