Verify that var promises don't contain more than one type definiton. (bug596)
It's possible for a variable promise to be defined with more than one datatype
constraint, e.g. string => "foo", int => "42". This causes the variable to get
the value last set, with no warning issued.

We treat this as an error by having the parser call CheckPromise (syntax.c)
upon each added constraint.