set promise kept class for files promise unless sub-nodes are repaired, closes #860
If we have a recursive delete promise on a folder with no sub-nodes, we previously
didn't set any context. Now, after going through sub-nodes using DepthSearch and we
don't use include_basedir, we check to see if anything was repaired, if not we call
cfPS to set a kept context.