Move CBUNDLESEQUENCE into GenericAgentConfig
This turned out to be a precondition step for an issue which ultimately turned out
to be impossible to fix. I still elected to keep the factoring and this requires
an apology. I often find myself wanting to pass parameters in GenericAgentConfig
and this injects the struct into more functions, so I am confident it will be
helpful in the future.