Encapsulate parsing state
Previously parsing state was exposed to all the code as a global variable
`P'. This led to (either inadvertent or concious) use of this variable in
various places of code. Split parser interface into external (parser.h) and
internal (parser_state.h) pieces and only expose internal state to the parser
and lexer themselves.