Add some Policy checks and tests
Let a policy error be an error of the sanity of policy files, e.g.
that a promise fails to specify  a required parameter. Let a syntax
error be an error resulting from parser failure, e.g. a promise
statement is not terminated by a semicolon. By this definition, the
parser is currently doing basic policy checking in addtion to
syntax checking.

This commit introduces policy checks as separate from syntax checks.
Policy checks may be done on a well-formed parse tree, i.e. a
struct Policy successfully returned from the parser. The motivation
for this is to be able to test static policy sanity separate from parsing.