Rename Constraint.isbody -> Constraint.references_body
This became confusing since a constraint relates to a body in
two ways. It may reference a body, i.e. the symbol of its Rval
is the name of a body defined elsewhere. Or it may be that it is
a body constraint, i.e. it's parent element is a body as opposed
to a promise