New option `--trust-key`/`-t` for `cf-key`.
This new option copies the given public key file
to the file name where `cf-serverd` would have stored
it if it had been gotten as the result of a successful
network exchange.

Such an option is intended to be used in fully-automated
installs, when there is a trusted way of deploying client
keys on the server and vice-versa.

  modified:   cf-key.c