Implement subset comparison functionality.
Implement functionality for determining whether the contents of a tree match a subset of contents found in
another tree. Can be used when a tree to be deleted is defined using a subset of matching content. Also can
be used to determine whether a given tree already exists as a subset of another tree, before tree insertion.