Fixed invalid memory read in Agent2Type
Calling Agent2Type with a string which is not in the CF_AGENTTYPES array
would cause the loop to run from 0 to (int)cf_notype - which is 16 right
now. However, the CF_AGENTTYPES array doesn't have that many entries.
It's not a valid enum value of the cfagenttype enum either.  This looks
like a copy & paste issues from Typename2Datatype.

The code should rather loop until the largest value in the cfagenttype
enum: cf_noagent.