Made Str2ServicePolicy yield correct enum value for "restart"
The 'for' loop terminated to early, it didn't check the entire array so
the function decided that 'restart' is not a valid string in which case
the function yielded the default value: cfsrv_start. So "start" and
"restart" mapped to the same enum value even though "restart" has a
dedicated value.

I suspect this is not intended, so let's fix it by adjusting the for
loop so that 'restart' gets mapped to cfsrv_restart.