Link libxml2, libpq and libmysqlclient to cf-agent, not to cf-promises
The following link sequence causes problems if system-provided libxml2 is not
binary compatible with the one CFEngine links against:

apache2 -> php -> xml.so -> libxml2.so.1
              \-> cfmod.so -> libpromises.so.1 -> libxml2.so.1

So in order to avoid conflicts between OS-provided and CFEngine-built libraries,
libxml2, libpq and libmysqlclient has been made dependencies of cf-agent, not
libpromises.so.

This change involved moving file promise functionality to the cf-agent, where it
logically belongs anyway.