The operands of a logical && or || shall be primary-expressions (MISRA 12.5)
"Primary expressions" are defined in ISO 9899:1990, section 6.3.1. Essentially they are either a single identifier,
or a constant, or a parenthesised expression. The effect of this rule is to require that if an operand is other than
a single identifier or constant then it must be parenthesised. Parentheses are important in this situation both for
readability of code and for ensuring that the behaviour is as the programmer intended. Where an expression consists
of either a sequence of only logical && or a sequence of only logical ||, extra parentheses are not required.