lastseen: initialize variable, don't propagate garbled digest
When calling Address2Hostkey(address, result) the result would only
be set if the address is known in the lastseen database. However,
some code did not check the return value of the fn() and still used
the result as if it were valid.

Now, just initialize the result to an empty string anyway.
(cherry picked from commit 71e739220ab50755c8c647a92176e575472488b9)

Conflicts:

	src/lastseen.c