pipes.c: we must keep pid of children even when fd > MAX_FD
For that, we'd better realloc() the CHILDREN array than let any of the
forked process become defunct.

This was observed in a server doing many parallel slow operations. The
number of FDs was excausted, in a way that newly forked processes would
no longer be tracked and wait()ed for.