Fix initialization of Promise struct in TestExpandPromise
This addresses Redmine #2021 - https://cfengine.com/dev/issues/2021

When the namespace member was added to the Promises_ structure in
commit 09613d808adffefd926cb686a9c53e252bb61e6 the initialization
of that char pointer was added to NewPromise, but omitted from
TestExpandPromise, which builds a standalone Promise for testing.

When the code attempted to copy the test promise, it called xstrdup
on the uninitialized pointer in pp.namespace, resulting in a core
dump. By initializing the namespace string in TestExpandPromise,
this problem is avoided.
(cherry picked from commit 32a8e2f3670d12b6fddf0c03f9fa9fcbc16412a0)