Add --full-check flag to cf-promises
Previously, cf-promises did full validation checks requiring a
runnable policy (with body common control) as input. It also did
certain checks to see that bundle/body references were accounted for.

Now, cf-promises will validate a partial policy file if this is given.
If it is given a runnable policy file, it will additionally validate
the integrity as before.

To enforce integrity checks, use the new --full-check flag. This is
also being used by other agents when running cf-promises for validation.

The new lighter partial checks makes it possible to use cf-promises as
a tool operating on incomplete policies.