Introduce "Immediate constraint" promise accessor functions
Lists of constraints in expanded promises contain both constraints specified in
promise and constraints merged in from referenced bodies.

Constraints specified in promise can not be guarded by classes, and hence don't
need evaluation context to be accessed. This commit adds accessors for such
constraints.

Note that it is a stop-gap measure, as it relies on caller only asking for
constraints specified in promise. Longer-term solution is to separate those two
types of constraints using type system.