Bug #1332 Acceptance tests for XML fail due to textual comparison of results.
Developed a program, with help from Mikhail, to convert an xml file into its c14n canonical
equivalent. Also, we expanded the acceptance testing structure to perform a canonical
diff on the xml files. Then transered all the xml acceptance tests that are passing
from tests/acceptance/10_files/11_xml_edits/staging/ to tests/acceptance/10_files/11_xml_edits/.
The two tests that remain in staging are failing, for unrelated reasons.

Note: Origionally I implemented this as a function, in evalfunction.c but due to conflicts with
differing libxml2 libraries, we decided to re-implement this as a standalone program for use only
with acceptance testing infrastructure. At some point, it may be beneficial to introduce this
program, as well as various other xml functions, into the CFEngine language.