Increase polling SLEEP_POLL_TIMEOUT from 10us to 10ms.
The idea is that if a signal takes time to be delivered, it's because the
system is under heavy load, so don't flood the process. Also a sysadm won't
see thousands of kill lines on his strace logs. The extra delay is
insignificant, this should happen rarely, only when the DB has a PID that is
not cleared.