Fix for HPUX
We have a stub for sendto in logging_test.c. In HPUX this function
returns int instead of ssize_t, therefore we need a separate stub
for HPUX.

Removing DEBUG definition

This was added to try to compensate for the missing symbol when compiling on HPUX.
I added the real symbol instead by including cf3globals.c to the compilation stage.

Include cf3globals.c only when compiling for HPUX