Wrong signature for mocked function
In logging_test.c we mock "sendto" to avoid sending data. In HPUX the compiler
is a little bit more strict than anticipated so it was complaining about the
type of the signature. It is interesting that this worked all over the place
and no other compiler complained even though it is evidently a failure.
(cherry picked from commit fbd070496208f87498558cf1bb71b4fddac3c892)