Don't register a thread-ID callback on Windows.
pthread_t on MinGW32 is a struct, and long is not large enough to hold the void*
within.

OpenSSL documentation states: If the application does not register such a callback
using CRYPTO_THREADID_set_callback(), then a default implementation is used - on
Windows and BeOS this uses the system's default thread identifying APIs

So not setting this callback at all will be safe.