Make sure cf-serverd exits cleanly on signal
Great parallelism introduced by commit 029775ea brought in the light obscure
behaviour, like deadlocking on signal instead of exiting. To remedy this with
the minimal amount of changes (this is a stable branch...) we simplify the
signal handler to only set the EXITNOW flag, which is checked in the mainloop.
Moreover, to make sure no databases get opened by threads just before exiting,
we close them and keep the mutexes locked while exiting. Dirty but works...