Replace hardcoded command paths in acceptance tests with $(G.cmd)
The "G" bundle provides full path names for various commands, in order
to eliminate the need to hard code pathnames in the acceptance tests.
Some older operating systems may have Perl installed in /usr/local/bin,
for example.

This commit updates all acceptance tests to make use of the $(G.cmd)
variables for each of the listed commands in order to eliminate
hardcoding of command paths in order to increase cross-platform
compatibility.