Ensure lvals in VarRefs are not composite for EvalContextVariableGet
This is a prerequisite for a new symbol table. Previously the lval
compontent of a VarRef that was passed to EvalContextVariableGet could
contain a dot, meaning that we had parse this out before looking up the
variable. This would then take presidence over the scope component of
the same VarRef.

This commit asserts that VarRefs passed to EvalContextVariableGet are
proper VarRefs.