Add logic-less templates (Mustache) to files promises, acceptance test
This adds two constraints to the files promise type: template_method and template_data.
The former may be either 'cfengine' or 'mustache' and defaults to 'cfengine'. It specifies
which templating engine should be used for rendering. The latter constraint takes a
container value and is required when using the 'mustache' template method.

Example:

files:
  "nginx.conf"
     edit_template => "nginx.mustache",
     template_method => "mustache",
     template_data => readjson("nginx.json", 4096);