Avoid expansion overflow in successful comparisons.
If two files are identical, there is no reason to run the DEBUG
report commands. The 07_packages/repositories.cf test caused the
expansion to overflow even though the test itself was successful.
Fixing it by using an extra class to delay the execution until we
know the outcome.

This doesn't prevent the expansion from overflowing if the test fails,
but in that case it doesn't really matter.