Revert "cf-agent --bootstrap fails on Debian systems, because sys.workdir/inputs is a symlink to /etc/cfengine3."
This reverts commit deeac31cce0ed001e329ba56214752a97342b205.

Since this code is also used in files promises, it can deal with
untrusted directories, not just bootstrap directories. In such
directories it is not safe to assume that symlinks can be followed,
since they could point to any file on the host, and we would open them
with root privileges.