No need to copy the text before the substitution.
The original line_buff starts with that text and we'll want it to
still start with it when we're done; so just munge what's after it.

It's also pointless to snprintf(line_buff, ...) twice without doing
anything in between that uses the intermediate state; and the size to
specify includes space for the '\0'.