Make buffer size match what snprintf()'s told.
RemoteSysLog()'s message[] buffer was declared with one size and
passed to snprintf() with another.  Use sizeof() in the snprintf()
call, to ensure it's definitely right, and declare it with the size
that was previously claimed.  Move the declaration closer to use, too,
so that we omit the array entirely when not used.

Also adjust another call to pass an actual buffer size, without
subtracting 1; snprintf() counts the final '\0' towards buffer size
(albeit leaving it out of the return).