In Debian's /etc/isssue, trim part after / in version.
Testing always lists its version with a /sid suffix, to distinguish it
from the name it'll have once it gets released; so it's currently
jessie/sid, which shall become jessie when released.  The noddy
command to parse /etc/issue and compare the result with sys.flavour
didn't take that into account; now it does.  The test failed on
testing, but now passes.