Add a test override for the $(sys.local_libdir) variable.
This is order to run tests in a test directory, while still specifying
the masterfiles root as the location of the standard library. Without
this, when a library policy file tries to include another file via
"inputs" it gets the path wrong because all inputs are relative to the
main policy file, not to the current file.